Use force_str and allow partial column parsing
TL;DR: Make sure to have a version of this fork in our git repository

Long version: once upon a time a developer published a version 0.99.1
of this library without any commit on our fork on github. Therefore,
when we moved from an old devpi to a new one, and when we looked at
our application's dependencies, we couldn't see the reason for the
fork.

So we upgrade the dependencies... but it failed! So we had to dig into
all the fork of fork (I skip the details here), and we ended
discovering the truth: we no longuer had the correct version on any
git repository.

The solution is in this commit:

* patch to use six properly to detect non-unicode string
* patch to keep the `parse_all_cols` parameter
* patch to fix minor issues here and there
* update version to `0.99+oscaro1` instead of a fake `0.99.1`

This is a very sad story, full of mistakes. I could say a lot more
about this library, but whatever. I don't have time for that.